refactor: popularityScore 타입 Double→Long 변환 + flush 타이밍 보정
- 부동소수점 커서 동등비교 오차 방지를 위해 Long 타입으로 변경
- 가중치 10배 스케일업 (reactions*30 + scraps*20 + epoch/8640)
- V25 마이그레이션 DOUBLE PRECISION → BIGINT, CAST 적용
- recalculatePopularityScore에 flushAutomatically=true 추가하여
  delete 직후 COUNT가 이전 값을 읽는 문제 방지

+++ src/main/java/com/therapyCommunity_Vol1/backend/post/dto/PopularCursor.java
@@ -7,7 +7,7 @@ import com.therapyCommunity_Vol1.backend.global.exception.ErrorCode;
 import java.util.Base64;
 
 public record PopularCursor(
-        Double score,
+        Long score,
         Long id
 ) {
     private static final ObjectMapper MAPPER = new ObjectMapper();
